refactor(api): 按领域拆分 api 模块
将 45 个 API 方法从单个 index.ts 拆分为按职责划分的模块：

  client.ts     — 基础 request 函数、token、共用类型
  sessions.ts   — 会话 CRUD、undo、constify
  providers.ts  — 供应商 CRUD、连接测试、模型发现
  memories.ts   — 叙事、瞬间、记忆分区
  personas.ts   — 人设读写
  settings.ts   — 路径白名单、拒止锚、环境变量
  system.ts     — 健康检查、重启、系统动态、DeepSeek 余额
  anthropic.ts  — Skills / Tools / Macros
  files.ts      — 文件选择器、图片服务

index.ts 保持扁平 api 对象导出，所有消费者无需改动导入路径。
