+++ web/src/api/providers.ts
@@ -1,0 +1,52 @@
+import type {
+  ProviderConfig,
+  ListProvidersResponse,
+  TestConnectionResponse,
+  DiscoverModelsResponse,
+} from '@/types'
+import { request, type ConnectionTestInput } from './client'
+
+export const providersApi = {
+  listProviders: () =>
+    request<ListProvidersResponse>('/providers'),
+
+  getProvider: (id: string) =>
+    request<ProviderConfig>(`/providers/${id}`),
+
+  createProvider: (body: Partial<ProviderConfig>) =>
+    request<ProviderConfig>('/providers', {
+      method: 'POST',
+      body: JSON.stringify(body),
+    }),
+
+  updateProvider: (id: string, body: Partial<ProviderConfig>) =>
+    request<ProviderConfig>(`/providers/${id}`, {
+      method: 'PUT',
+      body: JSON.stringify(body),
+    }),
+
+  deleteProvider: (id: string) =>
+    request<{ status: string }>(`/providers/${id}`, { method: 'DELETE' }),
+
+  testConnection: (body: ConnectionTestInput) =>
+    request<TestConnectionResponse>('/providers/test', {
+      method: 'POST',
+      body: JSON.stringify(body),
+    }),
+
+  discoverModels: (body: ConnectionTestInput) =>
+    request<DiscoverModelsResponse>('/providers/discover-models', {
+      method: 'POST',
+      body: JSON.stringify(body),
+    }),
+
+  discoverModelsForExisting: (id: string) =>
+    request<DiscoverModelsResponse>(`/providers/${id}/discover-models`, {
+      method: 'POST',
+    }),
+
+  testExistingProvider: (id: string) =>
+    request<TestConnectionResponse>(`/providers/${id}/test`, {
+      method: 'POST',
+    }),
+}
